# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * JobPilot application API
 * OpenAPI spec version: 0.1.0
 */
export interface ErrorResponse {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface PipelineStage {
  status: string;
  count: number;
}

export interface Dashboard {
  jobsDiscovered: number;
  queued: number;
  applied: number;
  interviews: number;
  responseRate: number;
  dailyUsed: number;
  dailyLimit: number;
  pipeline: PipelineStage[];
}

export type EvidenceItemCategory = typeof EvidenceItemCategory[keyof typeof EvidenceItemCategory];


export const EvidenceItemCategory = {
  experience: 'experience',
  achievement: 'achievement',
  education: 'education',
  certification: 'certification',
  project: 'project',
} as const;

export interface EvidenceItem {
  /** @minLength 1 */
  id: string;
  category: EvidenceItemCategory;
  /** @minLength 1 */
  title: string;
  organization: string;
  startDate?: string;
  endDate?: string;
  /** @minLength 1 */
  detail: string;
  /** @minLength 1 */
  source: string;
}

export interface Profile {
  id: number;
  fullName: string;
  email: string;
  phone?: string;
  location?: string;
  headline?: string;
  targetRoles: string[];
  preferredLocations: string[];
  workModes: string[];
  minimumSalary: number;
  skills: string[];
  evidence: EvidenceItem[];
  completion: number;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  fullName: string;
  email: string;
  phone?: string;
  location?: string;
  headline?: string;
  targetRoles: string[];
  preferredLocations: string[];
  workModes: string[];
  /** @minimum 0 */
  minimumSalary: number;
  skills: string[];
  evidence: EvidenceItem[];
}

export interface Job {
  id: number;
  company: string;
  title: string;
  location: string;
  workMode: string;
  salary: string;
  fitScore: number;
  status: string;
  source: string;
  sourceUrl?: string;
  postedAt: string;
  matchedSkills: string[];
  missingSkills: string[];
  description: string;
}

export interface JobInput {
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  title: string;
  location: string;
  workMode: string;
  salary: string;
  source: string;
  sourceUrl?: string;
  /** @minLength 1 */
  description: string;
}

export interface JobUpdate {
  status?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  fitScore?: number;
}

export type PrepareApplicationInputMode = typeof PrepareApplicationInputMode[keyof typeof PrepareApplicationInputMode];


export const PrepareApplicationInputMode = {
  review: 'review',
  batch: 'batch',
} as const;

export interface PrepareApplicationInput {
  mode: PrepareApplicationInputMode;
}

export interface Application {
  id: number;
  jobId: number;
  company: string;
  title: string;
  mode: string;
  portalType: string;
  status: string;
  progress: number;
  executionState: string;
  currentStep: string;
  portalUrl?: string;
  /** @nullable */
  blocker?: string | null;
  /** @nullable */
  lastError?: string | null;
  requiresUserAction: boolean;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  lastReplyAt?: string | null;
  /** @nullable */
  lastReplyType?: string | null;
  /** @nullable */
  lastReplySubject?: string | null;
  /** @nullable */
  lastReplySnippet?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ApplicationUpdate {
  status: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  /** @nullable */
  blocker?: string | null;
}

export interface ApplicationExecutionInput {
  submitAutomatically: boolean;
}

export interface ApplicationExecutionSubmitInput {
  /** @minLength 20 */
  approvalToken: string;
}

export interface ExecutionEvent {
  id: number;
  type: string;
  message: string;
  createdAt: string;
}

export interface ExecutionStatus {
  applicationId: number;
  portalType: string;
  executionState: string;
  currentStep: string;
  filledFields: number;
  totalFields: number;
  unresolvedFields: string[];
  requiresUserAction: boolean;
  /** @nullable */
  approvalToken: string | null;
  events: ExecutionEvent[];
}

export interface GenerationInput {
  /** @maxLength 2000 */
  instructions: string;
}

export type GenerationUpdateStatus = typeof GenerationUpdateStatus[keyof typeof GenerationUpdateStatus];


export const GenerationUpdateStatus = {
  draft: 'draft',
  approved: 'approved',
} as const;

export interface GenerationUpdate {
  cvMarkdown?: string;
  coverLetterMarkdown?: string;
  status?: GenerationUpdateStatus;
}

export interface GenerationClaim {
  section: string;
  text: string;
  /** @minItems 1 */
  evidenceIds: string[];
}

export interface Generation {
  id: number;
  jobId: number;
  version: number;
  status: string;
  model: string;
  cvMarkdown: string;
  coverLetterMarkdown: string;
  evidenceIds: string[];
  claims: GenerationClaim[];
  rejectedClaims: string[];
  createdAt: string;
  updatedAt: string;
}

export interface BatchSettings {
  enabled: boolean;
  dailyLimit: number;
  delayMinutes: number;
  minimumFitScore: number;
  requireReviewForSensitive: boolean;
  stopOnCaptcha: boolean;
  stopOnMfa: boolean;
}

export interface BatchSettingsUpdate {
  enabled: boolean;
  /**
     * @minimum 1
     * @maximum 100
     */
  dailyLimit: number;
  /**
     * @minimum 1
     * @maximum 120
     */
  delayMinutes: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  minimumFitScore: number;
  requireReviewForSensitive: boolean;
  stopOnCaptcha: boolean;
  stopOnMfa: boolean;
}

export interface BatchRunInput {
  /** @minItems 1 */
  jobIds: number[];
}

export interface BatchRun {
  id: number;
  status: string;
  queuedCount: number;
  startedAt: string;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  timestamp: string;
}

export interface TrackingSync {
  scannedThreads: number;
  matchedMessages: number;
  newMessages: number;
  updatedApplications: number;
}

/**
 * A valid Clerk session is required.
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * The authenticated request is not allowed.
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * The per-user AI request limit was reached.
 */
export type RateLimitedResponse = ErrorResponse;

export type ListJobsParams = {
status?: string;
search?: string;
};