# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JobPilot application API
servers:
  - url: /api
    description: Base API path
security:
  - clerkSession: []
  - bearerAuth: []
...
paths:
...
  /applications/{id}/execution/prepare:
    post:
      operationId: prepareApplicationExecution
      tags: [applications]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationExecutionInput"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Application not found }
        "200":
          description: Prepared application execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionStatus"
  /applications/{id}/execution:
    get:
      operationId: getApplicationExecution
      tags: [applications]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Application not found }
        "200":
          description: Application execution status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionStatus"
  /applications/{id}/execution/submit:
    post:
      operationId: submitApplicationExecution
      tags: [applications]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationExecutionSubmitInput"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Application not found }
        "409": { description: Execution session is missing or expired }
        "200":
          description: Submitted application execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionStatus"
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk browser session cookie.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token for non-browser clients.
  responses:
    Unauthorized:
      description: A valid Clerk session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated request is not allowed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ApplicationExecutionInput:
      type: object
      required: [submitAutomatically]
      properties:
        submitAutomatically: { type: boolean }
    ApplicationExecutionSubmitInput:
      type: object
      required: [approvalToken]
      properties:
        approvalToken: { type: string, minLength: 20 }
    ExecutionEvent:
      type: object
      required: [id, type, message, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        message: { type: string }
        createdAt: { type: string }
    ExecutionStatus:
      type: object
      required: [applicationId, portalType, executionState, currentStep, filledFields, totalFields, unresolvedFields, requiresUserAction, approvalToken, events]
      properties:
        applicationId: { type: integer }
        portalType: { type: string }
        executionState: { type: string }
        currentStep: { type: string }
        filledFields: { type: integer }
        totalFields: { type: integer }
        unresolvedFields:
          type: array
          items: { type: string }
        requiresUserAction: { type: boolean }
        approvalToken: { type: ["string", "null"] }
        events:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionEvent"
...