# lib/api-zod/src/generated/types/applicationExecutionSubmitInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * JobPilot application API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationExecutionSubmitInput {
  /** @minLength 20 */
  approvalToken: string;
}